refactor: refactored helper print and removed not needed logger function style: cs fixes feat: added more docs
