+++ packages/cerebro-core/docs/toolbox-meta.md
@@ -1,0 +1,50 @@
+Provide functions for accessing information about the currently running CLI.
+
+You can access these tools on the cerebro toolbox, via `const { toolbox } = require('@anolilab/cerebro-core')`.
+
+## src
+
+The currently running CLI's source folder.
+
+```js
+toolbox.meta.src // "/Users/private/cerebro"
+```
+
+## version
+
+Retrieves the currently running CLI's version.
+
+```js
+toolbox.meta.version() // '1.0.0'
+```
+
+## packageJSON
+
+Retrieves the currently running CLI's package.json contents as an object.
+
+```js
+toolbox.meta.packageJSON()
+// { name: 'cerebro', version: '1.0.0', ... }
+```
+
+## checkForUpdate
+
+Async function that checks NPM to see if there's an update to the currently running CLI.
+
+```js
+const newVersion = await toolbox.meta.checkForUpdate()
+// false (if none exists)
+// '1.0.0' (new version if exists)
+if (newVersion) {
+  toolbox.print.info(`New version available: ${newVersion})`)
+}
+```
+
+## commandInfo
+
+Retrieves information about all of this CLI's commands. You can use this to display a custom help screen, for example.
+
+```js
+const commandInfo = toolbox.meta.commandInfo()
+toolbox.print.table(commandInfo)
+```
