+++ packages/cerebro-core/docs/toolbox-print.md
@@ -1,0 +1,139 @@
+Features for allowing you to print to the console.
+
+You can access these tools on the cerebro toolbox, via `const { toolbox } = require('@anolilab/cerebro-core')`.
+
+## colors
+
+An object for working with printing colors on the command line. It is from the `colors` NPM package,
+however we define a theme to make things a bit consistent.
+
+Some available functions include:
+
+| function             | use when you want...                     |
+| -------------------- | ---------------------------------------- |
+| `colors.success()`   | the user to smile                        |
+| `colors.error()`     | to say something has failed              |
+| `colors.warning()`   | to point out that something might be off |
+| `colors.highlight()` | to draw attention to something           |
+| `colors.info()`      | to say something informational           |
+| `colors.muted()`     | you need to say something secondary      |
+
+Each take a `string` parameter and return a `string`.
+
+One gotcha here is that the length of the string is longer than you think because of the embedded
+color codes that disappear when you print them.
+
+## spin
+
+Creates a spinner for long running tasks on the command line. It's
+[ora](https://github.com/sindresorhus/ora)!
+
+Here's an example of how to work with it:
+
+```js
+// a spinner starts with the text you provide
+const spinner = toolbox.print.spin('Time for fun!')
+await toolbox.system.run('sleep 5')
+```
+
+> Important: make sure you don't print anything else while a spinner is going. You need to stop
+it first.
+
+There's a few ways to stop it.
+
+```js
+// stop it & clear the text
+spinner.stop()
+
+// stop it, leave a checkmark, and optional new text
+spinner.succeed('woot!')
+
+// stop it, leave an X, and optional new text
+spinner.fail('womp womp.')
+
+// stop it, leave a custom label, and optional new text
+spinner.stopAndPersist({ symbol: '🚨', text: 'osnap!' })
+```
+
+Once stopped, you can start it again later.
+
+```js
+spinner.start()
+```
+
+You can change the color of the spinner by setting:
+
+```js
+spinner.color = 'cyan'
+```
+
+The text can also be set with the normal printing colors.
+
+```js
+spinner.text = toolbox.print.colors.green('i like trees')
+```
+
+## table
+
+Prints out a table of data, including a header. You can choose from three different formats:
+`default`, `markdown`, and `lean`.
+
+```js
+const { table } = toolbox.print
+table(
+  [
+    ['First Name', 'Last Name', 'Age'],
+    ['Jamon', 'Holmgren', 35],
+    ['Gant', 'Laborde', 36],
+    ['Steve', 'Kellock', 43],
+    ['Gary', 'Busey', 73],
+  ],
+  { format: 'markdown' },
+)
+```
+
+Output:
+
+```
+| First Name | Last Name | Age |
+| ---------- | --------- | --- |
+| Jamon      | Holmgren  | 35  |
+| Gant       | Laborde   | 36  |
+| Steve      | Kellock   | 43  |
+| Gary       | Busey     | 73  |
+```
+
+You can also pass styles for the table (as specified in [cli-table3](https://github.com/cli-table/cli-table3)):
+
+```js
+const { table } = toolbox.print
+table(
+  [
+    ['First Name', 'Last Name', 'Age'],
+    ['Jamon', 'Holmgren', 35],
+    ['Gant', 'Laborde', 36],
+    ['Steve', 'Kellock', 43],
+    ['Gary', 'Busey', 73],
+  ],
+  {
+    format: 'lean',
+    style: { 'padding-left': 0 , 'padding-right': 8 }
+  },
+)
+```
+
+Output:
+
+```
+┌──────────────────┬─────────────────┬───────────┐
+│First Name        │Last Name        │Age        │
+├──────────────────┼─────────────────┼───────────┤
+│Jamon             │Holmgren         │35         │
+├──────────────────┼─────────────────┼───────────┤
+│Gant              │Laborde          │36         │
+├──────────────────┼─────────────────┼───────────┤
+│Steve             │Kellock          │43         │
+├──────────────────┼─────────────────┼───────────┤
+│Gary              │Busey            │73         │
+└──────────────────┴─────────────────┴───────────┘
+```
