+++ packages/cerebro-core/src/index.ts
@@ -71,7 +71,6 @@ export type {
     PackageJSON,
     Logger,
     Loader,
-    ConfigType,
     Extension,
     ExtensionSetup,
     ExtensionOverrides,
